+++ pytorch_segmentation_models_trainer/custom_metrics/metrics.py
@@ -16,21 +16,21 @@
  *   the Free Software Foundation; either version 2 of the License, or     *
  *   (at your option) any later version.                                   *
  *                                                                         *
- *   Code inspired by the one in                                           *
- *   https://github.com/Lydorn/Polygonization-by-Frame-Field-Learning/     *
  ****
 """
 
 import itertools
 from typing import List, Tuple, Union
 import numpy as np
+import similaritymeasures as sm
 from shapely.geometry import Polygon, LineString, Point
+from shapely.geometry.base import BaseGeometry
 from shapely.geometry.multipolygon import MultiPolygon
 import torch
 from pytorch_segmentation_models_trainer.utils import polygonrnn_utils
 
 
-def iou(y_pred, y_true, threshold):
+def iou(y_pred: torch.Tensor, y_true: torch.Tensor, threshold: float) -> float:
     assert (
         len(y_pred.shape) == len(y_true.shape) == 2
     ), "Input tensor shapes should be (N, .)"
@@ -43,7 +43,9 @@ def iou(y_pred, y_true, threshold):
     return r
 
 
-def polygon_iou(vertices1: List, vertices2: List) -> Tuple[float, float, float]:
+def polygon_iou(
+    vertices1: Union[List, BaseGeometry], vertices2: Union[List, BaseGeometry]
+) -> Tuple[float, float, float]:
     """
     calculate iou of two polygons
     :param vertices1: vertices of the first polygon
@@ -55,6 +57,20 @@ def polygon_iou(vertices1: List, vertices2: List) -> Tuple[float, float, float]:
     if geom1.is_valid and geom2.is_valid:
         return _poly_iou(geom1, geom2)
     return _iou_with_invalid_geom(geom1, geom2)
+
+
+def shapely_polygon_iou(polygon_a: Polygon, polygon_b: Polygon) -> float:
+    """Compute the iou between two polygons.
+
+    Args:
+        polygon_a (Polygon): Shapely polygon
+        polygon_b (Polygon): Shapely polygon
+
+    Returns:
+        float: iou
+    """
+    iou, _, __ = polygon_iou(polygon_a, polygon_b)
+    return iou
 
 
 def _iou_with_invalid_geom(
@@ -107,7 +123,7 @@ def polis(polygon_a: Polygon, polygon_b: Polygon) -> float:
     )
 
 
-def _one_side_polis(coords: List, bounds: LineString) -> float:
+def _one_side_polis(coords: List[Tuple[float, float]], bounds: LineString) -> float:
     """Compute the polis metric for one side of a polygon.
 
     Args:
@@ -123,7 +139,10 @@ def _one_side_polis(coords: List, bounds: LineString) -> float:
     return float(distance_sum / float(2 * len(coords)))
 
 
-def batch_polis(batch_polygon_a: np.ndarray, batch_polygon_b: np.ndarray) -> np.ndarray:
+def batch_polis(
+    batch_polygon_a: Union[np.ndarray, List[np.ndarray]],
+    batch_polygon_b: Union[np.ndarray, List[np.ndarray]],
+) -> np.ndarray:
     """Compute the polis metric between two polygon batches.
 
     Args:
@@ -134,5 +153,30 @@ def batch_polis(batch_polygon_a: np.ndarray, batch_polygon_b: np.ndarray) -> np.
             return 0
         return polis(Polygon(numpy_polygon_a), Polygon(numpy_polygon_b))
 
-    func = lambda x: _polis(x[0], x[1])
+    def func(x):
+        return _polis(x[0], x[1])
+
     return np.array(list(map(func, zip(batch_polygon_a, batch_polygon_b))))
+
+
+def hausdorff_distance(geom1: BaseGeometry, geom2: BaseGeometry) -> float:
+    """Compute the hausdorf distance between two geometries.
+
+    Args:
+        geom1 (BaseGeometry): Shapely geometry
+        geom2 (BaseGeometry): Shapely geometry
+
+    Returns:
+        float: hausdorf distance
+    """
+    return geom1.hausdorff_distance(geom2)
+
+
+def frechet_distance(
+    polygon1: Polygon, polygon2: Polygon, minkowski_p_norm: float = 2
+) -> float:
+    return sm.frechet_dist(
+        np.array(polygon1.exterior.coords),
+        np.array(polygon2.exterior.coords),
+        p=minkowski_p_norm,
+    )