+++ pytorch_segmentation_models_trainer/custom_models/rnn/polygon_rnn.py
@@ -24,6 +24,7 @@ from collections import OrderedDict
 from logging import log
 import os
 from pathlib import Path
+from typing import Dict
 
 import segmentation_models_pytorch as smp
 import torch
@@ -249,6 +250,18 @@ class ConvLSTM(nn.Module):
         return param
 
 
+def make_basic_conv_block(input_size, output_size, kernel_size, stride, padding):
+    """
+
+    :rtype: nn.Sequential
+    """
+    return nn.Sequential(
+        nn.Conv2d(input_size, output_size, kernel_size, stride, padding),
+        nn.ReLU(),
+        nn.BatchNorm2d(output_size),
+    )
+
+
 class PolygonRNN(nn.Module):
     """
     Code extracted from https://github.com/AlexMa011/pytorch-polygon-rnn
@@ -256,18 +269,6 @@ class PolygonRNN(nn.Module):
 
     def __init__(self, load_vgg=True, encoder_trainable=True, grid_size=28):
         super(PolygonRNN, self).__init__()
-
-        def _make_basic(input_size, output_size, kernel_size, stride, padding):
-            """
-
-            :rtype: nn.Sequential
-            """
-            return nn.Sequential(
-                nn.Conv2d(input_size, output_size, kernel_size, stride, padding),
-                nn.ReLU(),
-                nn.BatchNorm2d(output_size),
-            )
-
         self.grid_size = grid_size
         self.model1 = nn.Sequential(
             nn.Conv2d(3, 64, 3, 1, 1),
@@ -320,11 +321,11 @@ class PolygonRNN(nn.Module):
             nn.BatchNorm2d(512),
             nn.ReLU(),
         )
-        self.convlayer1 = _make_basic(128, 128, 3, 1, 1)
-        self.convlayer2 = _make_basic(256, 128, 3, 1, 1)
-        self.convlayer3 = _make_basic(512, 128, 3, 1, 1)
-        self.convlayer4 = _make_basic(512, 128, 3, 1, 1)
-        self.convlayer5 = _make_basic(512, 128, 3, 1, 1)
+        self.convlayer1 = make_basic_conv_block(128, 128, 3, 1, 1)
+        self.convlayer2 = make_basic_conv_block(256, 128, 3, 1, 1)
+        self.convlayer3 = make_basic_conv_block(512, 128, 3, 1, 1)
+        self.convlayer4 = make_basic_conv_block(512, 128, 3, 1, 1)
+        self.convlayer5 = make_basic_conv_block(512, 128, 3, 1, 1)
         self.poollayer = nn.MaxPool2d(2, 2)
         self.upsample = nn.Upsample(scale_factor=2, mode="bilinear")
         self.convlstm = ConvLSTM(
@@ -568,7 +569,9 @@ class PolygonRNN(nn.Module):
 
         return result
 
-    def compute(self, batch: torch.Tensor, image_key=None) -> torch.Tensor:
+    def compute(
+        self, batch: Dict[str, torch.Tensor], image_key: str = None
+    ) -> torch.Tensor:
         image_key = "image" if image_key is None else image_key
         output = self.forward(batch[image_key], batch["x1"], batch["x2"], batch["x3"])
         return output.contiguous().view(-1, self.grid_size * self.grid_size + 3)