+++ pytorch_segmentation_models_trainer/model_loader/detection_model.py
@@ -19,18 +19,10 @@
  ****
 """
 
-from pytorch_segmentation_models_trainer.model_loader.model import Model
-from torch.utils.data import DataLoader
 import torch
-from torchvision.ops import box_iou
-
-
-def _evaluate_iou(target, pred):
-    """Evaluate intersection over union (IOU) for target from dataset and output prediction from model."""
-    if pred["boxes"].shape[0] == 0:
-        # no box detected, 0 IOU
-        return torch.tensor(0.0, device=pred["boxes"].device)
-    return box_iou(target["boxes"], pred["boxes"]).diag().mean()
+from pytorch_segmentation_models_trainer.model_loader.model import Model
+from pytorch_segmentation_models_trainer.utils import object_detection_utils
+from torch.utils.data import DataLoader
 
 
 class ObjectDetectionPLModel(Model):
@@ -83,7 +75,7 @@ class ObjectDetectionPLModel(Model):
         )
 
     def training_step(self, batch, batch_idx):
-        images, targets = batch
+        images, targets, _ = batch
         loss_dict = self.model(images, targets)
         return {
             "loss": sum(loss for loss in loss_dict.values()),
@@ -92,12 +84,17 @@ class ObjectDetectionPLModel(Model):
         }
 
     def validation_step(self, batch, batch_idx):
-        images, targets = batch
+        images, targets, _ = batch
         self.model.train()
         loss_dict = self.model(images, targets)
         self.model.eval()
         outs = self.model(images)
-        iou = torch.stack([_evaluate_iou(t, o) for t, o in zip(targets, outs)]).mean()
+        iou = torch.stack(
+            [
+                object_detection_utils.evaluate_box_iou(t, o)
+                for t, o in zip(targets, outs)
+            ]
+        ).mean()
         return {
             "val_iou": iou,
             "loss": sum(loss for loss in loss_dict.values()),