+++ tests/test_tensor_utils.py
@@ -42,7 +42,7 @@ current_dir = os.path.dirname(__file__)
 root_dir = os.path.join(current_dir, "testing_data")
 
 
-class Test_TestTensorUtils(unittest.TestCase):
+class Test_TensorUtils(unittest.TestCase):
     def setUp(self):
         warnings.simplefilter("ignore", category=ImportWarning)
         warnings.simplefilter("ignore", category=DeprecationWarning)