+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# Version 0.16.0
+
+- New Mod PolyMapper model;
+- Matching methods added;
+- Evaluation methods added;
+
 # Veresion 0.15.0
 
 - New Naive Mod PolyMapper model (Object Detection + PolygonRNN);